[WFSSL-63] Update tests to no longer make use of TLSv1 and TLSv1.1 where possible. For tests where these protocols are needed, ensure that they are re-enabled if necessary
