+++ java/pom.xml
@@ -116,11 +116,10 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
-                <version>2.17</version>
 
                 <configuration>
                     <redirectTestOutputToFile>${maven.test.redirectTestOutputToFile}</redirectTestOutputToFile>
-                    <forkMode>once</forkMode>
+                    <reuseForks>false</reuseForks>
                     <systemProperties>
                         <javax.net.ssl.keyStore>src/test/resources/client.keystore</javax.net.ssl.keyStore>
                         <javax.net.ssl.trustStore>src/test/resources/client.truststore</javax.net.ssl.trustStore>
